backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
openai==1.74.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.1.0
tiktoken==0.9.0 
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1
redis==5.0.1
aiohttp==3.9.3
tenacity==8.2.3
pytest==8.3.5
pytest-asyncio==0.24.0
loguru==0.7.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client 